requirement.txt
requests
ccxt
pandas
numpy
matplotlib
twilio
ta
pytz
python-telegram-bot
configparser
openai==0.28